test(e2e): consolidate & dedup smoke specs (phases 1-4)
- Merge public-pages-extended into a parameterized public-pages.spec.ts
  (preserves >=28 team-row + >=1 row checks); delete the extended file.
- Delete auth-pages.spec.ts (zero assertNoPhpErrors); fold the unique
  DepthChartEntry "no Sign In + form visible" smoke into auth-pages-extended.
- Merge error-pages + modules-dispatch into module-routing.spec.ts (one
  canonical unknown-module-redirect test; both path-traversal encodings kept);
  drop the duplicate redirect test from flows/parameter-edge-cases.spec.ts.
- Remove the redundant olympics-coverage loop test; document the unique
  header/conference assertions live in smoke/olympics-pages.spec.ts.
- Replace mobile-subpages body.length>100 non-assertions with real anchor
  visibility checks (.ibl-data-table / .team-page-layout).

Verified green against the CI-seeded worktree DB (50 specs across the touched
files pass, incl. the row-count assertions).

Nightly note: bin/check-plan-staleness flagged the to-be-created
module-routing.spec.ts as a false positive (bare -e existence check cannot
distinguish a create-target from a deleted path); all source files resolve.
Override is intentional, not a stale plan.
